functions/function-4/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud
google-cloud-bigquery
google-cloud-storage
pandas
datetime
gcsfs
fsspec